app/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
psycopg2-binary==2.9.9
pydantic==2.4.2